+++ shoop/themes/classic_gray/plugins/__init__.py
@@ -1,0 +1,16 @@
+# This file is part of Shoop.
+#
+# Copyright (c) 2012-2016, Shoop Ltd. All rights reserved.
+#
+# This source code is licensed under the AGPLv3 license found in the
+# LICENSE file in the root directory of this source tree.
+
+from .product_highlight import ProductHighlightPlugin
+from .snippets import SnippetsPlugin
+from .social_media_links import SocialMediaLinksPlugin
+
+__all__ = [
+    "ProductHighlightPlugin",
+    "SnippetsPlugin",
+    "SocialMediaLinksPlugin",
+]